Rename PostsPageTitle.tsx to PostsPageHeaderTitle.tsx
